+++ ingeniamotion/fsoe.py
@@ -1,5 +1,7 @@
-import time
-from typing import TYPE_CHECKING, Dict, Optional
+import threading
+from dataclasses import dataclass
+from functools import partial
+from typing import TYPE_CHECKING, Callable, Dict, List, Optional
 
 import ingenialogger
 
@@ -8,10 +10,15 @@ try:
         ApplicationParameter,
         Dictionary,
         DictionaryItem,
+        DictionaryItemInput,
         DictionaryItemInputOutput,
         MasterHandler,
         StateData,
     )
+
+    if TYPE_CHECKING:
+        from fsoe_master.fsoe_master import State
+
 except ImportError:
     FSOE_MASTER_INSTALLED = False
 else:
@@ -20,12 +27,23 @@ else:
 from ingenialink.enums.register import REG_ACCESS, REG_DTYPE
 from ingenialink.ethercat.register import EthercatRegister
 from ingenialink.pdo import RPDOMap, RPDOMapItem, TPDOMap, TPDOMapItem
+from ingenialink.utils._utils import dtype_value
 
+from ingeniamotion.enums import FSoEState
 from ingeniamotion.exceptions import IMTimeoutError
 from ingeniamotion.metaclass import DEFAULT_SERVO
 
 if TYPE_CHECKING:
     from ingeniamotion.motion_controller import MotionController
+
+
+@dataclass
+class FSoEError:
+    """FSoE Error descriptor"""
+
+    servo: str
+    transition_name: str
+    description: str
 
 
 class FSoEMasterHandler:
@@ -40,9 +58,22 @@ class FSoEMasterHandler:
 
     STO_COMMAND_KEY = 0x040
     STO_COMMAND_UID = "STO_COMMAND"
+    SS1_COMMAND_KEY = 0x050
+    SS1_COMMAND_UID = "SS1_COMMAND"
+    SBC_COMMAND_KEY = 0x060
+    SBC_COMMAND_UID = "SBC_COMMAND"
+    SAFE_INPUTS_KEY = 0x070
+    SAFE_INPUTS_UID = "SAFE_INPUTS"
     PROCESS_DATA_COMMAND = 0x36
 
-    def __init__(self, slave_address: int, connection_id: int, watchdog_timeout: float):
+    def __init__(
+        self,
+        slave_address: int,
+        connection_id: int,
+        watchdog_timeout: float,
+        application_parameters: List["ApplicationParameter"],
+        report_error_callback: Callable[[str, str], None],
+    ):
         if not FSOE_MASTER_INSTALLED:
             return
         self.__master_handler = MasterHandler(
@@ -50,24 +81,32 @@ class FSoEMasterHandler:
             slave_address=slave_address,
             connection_id=connection_id,
             watchdog_timeout_s=watchdog_timeout,
-            application_parameters=[
-                ApplicationParameter(name="SAFETY_ADDRESS", initial_value=slave_address, n_bytes=2)
-            ],
+            application_parameters=application_parameters,
+            report_error_callback=report_error_callback,
+            state_change_callback=self.__state_change_callback,
         )
         self._configure_master()
         self.__safety_master_pdu = RPDOMap()
         self.__safety_slave_pdu = TPDOMap()
         self._configure_pdo_maps()
         self.__running = False
+        self.__state_is_data = threading.Event()
+
+        # The saco slave might take a while to answer with a valid command
+        # During it's initialization it will respond with 0's, that are ignored
+        # To avoid triggering additional errors
+        self.__in_initial_reset = False
 
     def _start(self) -> None:
         """Start the FSoE Master handler."""
+        self.__in_initial_reset = True
         self.__master_handler.start()
         self.__running = True
 
     def stop(self) -> None:
         """Stop the master handler"""
         self.__master_handler.stop()
+        self.__in_initial_reset = False
         self.__running = False
 
     def delete(self) -> None:
@@ -77,10 +116,6 @@ class FSoEMasterHandler:
     def _configure_pdo_maps(self) -> None:
         """Configure the PDOMaps used for the Safety PDUs."""
         PDUMapper.configure_rpdo_map(self.safety_master_pdu_map)
-        # Set the default initial value to the Safety Master PDU PDOMap
-        self.safety_master_pdu_map.set_item_bytes(
-            int(0).to_bytes(self.safety_master_pdu_map.data_length_bytes, "little")
-        )
         PDUMapper.configure_tpdo_map(self.safety_slave_pdu_map)
 
     def _configure_master(self) -> None:
@@ -92,13 +127,20 @@ class FSoEMasterHandler:
         """Configure the FSoE master handler's SafeOutputs."""
         # Phase 1 mapping
         self.__master_handler.master.dictionary_map.add_by_key(self.STO_COMMAND_KEY, bits=1)
+        self.__master_handler.master.dictionary_map.add_by_key(self.SS1_COMMAND_KEY, bits=1)
+        self.__master_handler.master.dictionary_map.add_padding(bits=6)
+        self.__master_handler.master.dictionary_map.add_by_key(self.SBC_COMMAND_KEY, bits=1)
         self.__master_handler.master.dictionary_map.add_padding(bits=7)
 
     def _map_inputs(self) -> None:
         """Configure the FSoE master handler's SafeInputs."""
         # Phase 1 mapping
         self.__master_handler.slave.dictionary_map.add_by_key(self.STO_COMMAND_KEY, bits=1)
-        self.__master_handler.slave.dictionary_map.add_padding(bits=7)
+        self.__master_handler.slave.dictionary_map.add_by_key(self.SS1_COMMAND_KEY, bits=1)
+        self.__master_handler.slave.dictionary_map.add_padding(bits=6)
+        self.__master_handler.slave.dictionary_map.add_by_key(self.SBC_COMMAND_KEY, bits=1)
+        self.__master_handler.slave.dictionary_map.add_by_key(self.SAFE_INPUTS_KEY, bits=1)
+        self.__master_handler.slave.dictionary_map.add_padding(bits=6)
 
     def get_request(self) -> None:
         """Set the FSoE master handler request to the Safety Master PDU PDOMap"""
@@ -109,7 +151,16 @@ class FSoEMasterHandler:
     def set_reply(self) -> None:
         """Get the FSoE slave response from the Safety Slave PDU PDOMap and set it
         to the FSoE master handler."""
-        self.__master_handler.set_reply(self.safety_slave_pdu_map.get_item_bytes())
+        reply = self.safety_slave_pdu_map.get_item_bytes()
+        if self.__in_initial_reset:
+            if reply[0] == 0:
+                # Byte 0 of FSoE frame should always be the command
+                # 0 is not a valid command
+                return
+            else:
+                self.__in_initial_reset = False
+
+        self.__master_handler.set_reply(reply)
 
     def sto_deactivate(self) -> None:
         """Set the STO command to deactivate the STO"""
@@ -119,6 +170,22 @@ class FSoEMasterHandler:
     def sto_activate(self) -> None:
         """Set the STO command to activate the STO"""
         self.__master_handler.dictionary.set(self.STO_COMMAND_UID, False)
+
+    def ss1_deactivate(self) -> None:
+        """Set the SS1 command to deactivate the SS1"""
+        self.__master_handler.set_fail_safe(False)
+        self.__master_handler.dictionary.set(self.SS1_COMMAND_UID, True)
+
+    def ss1_activate(self) -> None:
+        """Set the SS1 command to activate the SS1"""
+        self.__master_handler.dictionary.set(self.SS1_COMMAND_UID, False)
+
+    def safe_inputs_value(self) -> bool:
+        """Get the safe inputs register value"""
+        safe_inputs_value = self.__master_handler.dictionary.get(self.SAFE_INPUTS_UID)
+        if not isinstance(safe_inputs_value, bool):
+            raise ValueError(f"Wrong value type. Expected type bool, got {type(safe_inputs_value)}")
+        return safe_inputs_value
 
     def is_sto_active(self) -> bool:
         """Check the STO state.
@@ -132,6 +199,12 @@ class FSoEMasterHandler:
             raise ValueError(f"Wrong value type. Expected type bool, got {type(sto_command)}")
         return sto_command
 
+    def __state_change_callback(self, state: "State") -> None:
+        if state == StateData:
+            self.__state_is_data.set()
+        else:
+            self.__state_is_data.clear()
+
     def wait_for_data_state(self, timeout: Optional[float] = None) -> None:
         """Wait the FSoE master handler to reach the Data state.
 
@@ -144,23 +217,43 @@ class FSoEMasterHandler:
             IMTimeoutError: If the Data state is not reached within the timeout.
 
         """
-        state_reached = False
-        init_time = time.time()
-        while not state_reached:
-            state_reached = self.__master_handler.state == StateData
-            if timeout and (init_time + timeout) < time.time():
-                raise IMTimeoutError("The FSoE Master did not reach the Data state")
+        if self.__state_is_data.wait(timeout=timeout) is False:
+            raise IMTimeoutError("The FSoE Master did not reach the Data state")
 
-    @staticmethod
-    def _saco_phase_1_dictionary() -> "Dictionary":
+    def _saco_phase_1_dictionary(self) -> "Dictionary":
         """Get the SaCo phase 1 dictionary instance"""
         sto_command_dict_item = DictionaryItemInputOutput(
-            key=0x040,
-            name="STO_COMMAND",
+            key=self.STO_COMMAND_KEY,
+            name=self.STO_COMMAND_UID,
             data_type=DictionaryItem.DataTypes.BOOL,
             fail_safe_input_value=True,
         )
-        return Dictionary([sto_command_dict_item])
+        ss1_command_dict_item = DictionaryItemInputOutput(
+            key=self.SS1_COMMAND_KEY,
+            name=self.SS1_COMMAND_UID,
+            data_type=DictionaryItem.DataTypes.BOOL,
+            fail_safe_input_value=True,
+        )
+        sbc_command_dict_item = DictionaryItemInputOutput(
+            key=self.SBC_COMMAND_KEY,
+            name=self.SBC_COMMAND_UID,
+            data_type=DictionaryItem.DataTypes.BOOL,
+            fail_safe_input_value=False,
+        )
+        safe_input_dict_item = DictionaryItemInput(
+            key=self.SAFE_INPUTS_KEY,
+            name=self.SAFE_INPUTS_UID,
+            data_type=DictionaryItem.DataTypes.BOOL,
+            fail_safe_value=False,
+        )
+        return Dictionary(
+            [
+                sto_command_dict_item,
+                ss1_command_dict_item,
+                sbc_command_dict_item,
+                safe_input_dict_item,
+            ]
+        )
 
     @property
     def safety_master_pdu_map(self) -> RPDOMap:
@@ -172,6 +265,16 @@ class FSoEMasterHandler:
         """The PDOMap used for the Safety Slave PDU."""
         return self.__safety_slave_pdu
 
+    @property
+    def state(self) -> FSoEState:
+        """Get the FSoE master state."""
+        return FSoEState(self.__master_handler.state.id)
+
+    @property
+    def running(self) -> bool:
+        """True if FSoE Master is started, else False"""
+        return self.__running
+
 
 class PDUMapper:
     """Helper class to configure the Safety PDU PDOMaps."""
@@ -182,9 +285,14 @@ class PDUMapper:
     # Phase 1 mapping
     FSOE_COMMAND_SIZE_BITS = 8
     STO_COMMAND_SIZE_BITS = 1
-    STO_COMMAND_PADDING_SIZE_BITS = 7
+    SS1_COMMAND_SIZE_BITS = 1
+    STO_COMMAND_PADDING_SIZE_BITS = 6
+    SBC_COMMAND_SIZE_BITS = 1
+    SBC_COMMAND_PADDING_SIZE_BITS = 7
     CRC_O_SIZE_BITS = 16
     CONN_ID_SIZE_BITS = 16
+    SAFE_INPUTS_SIZE_BITS = 1
+    SAFE_INPUTS_PADDING_SIZE_BITS = 6
 
     @classmethod
     def configure_rpdo_map(cls, rpdo_map: RPDOMap) -> None:
@@ -199,8 +307,14 @@ class PDUMapper:
         rpdo_map.add_item(fsoe_command_item)
         sto_command_item = RPDOMapItem(size_bits=cls.STO_COMMAND_SIZE_BITS)
         rpdo_map.add_item(sto_command_item)
-        padding_item = RPDOMapItem(size_bits=cls.STO_COMMAND_PADDING_SIZE_BITS)
-        rpdo_map.add_item(padding_item)
+        ss1_command_item = RPDOMapItem(size_bits=cls.SS1_COMMAND_SIZE_BITS)
+        rpdo_map.add_item(ss1_command_item)
+        sto_padding_item = RPDOMapItem(size_bits=cls.STO_COMMAND_PADDING_SIZE_BITS)
+        rpdo_map.add_item(sto_padding_item)
+        sbc_command_item = RPDOMapItem(size_bits=cls.SBC_COMMAND_SIZE_BITS)
+        rpdo_map.add_item(sbc_command_item)
+        sbc_padding_item = RPDOMapItem(size_bits=cls.SBC_COMMAND_PADDING_SIZE_BITS)
+        rpdo_map.add_item(sbc_padding_item)
         crc_0_item = RPDOMapItem(size_bits=cls.CRC_O_SIZE_BITS)
         rpdo_map.add_item(crc_0_item)
         conn_id_item = RPDOMapItem(size_bits=cls.CONN_ID_SIZE_BITS)
@@ -219,8 +333,16 @@ class PDUMapper:
         tpdo_map.add_item(fsoe_command_item)
         sto_command_item = TPDOMapItem(size_bits=cls.STO_COMMAND_SIZE_BITS)
         tpdo_map.add_item(sto_command_item)
-        padding_item = TPDOMapItem(size_bits=cls.STO_COMMAND_PADDING_SIZE_BITS)
-        tpdo_map.add_item(padding_item)
+        ss1_command_item = TPDOMapItem(size_bits=cls.SS1_COMMAND_SIZE_BITS)
+        tpdo_map.add_item(ss1_command_item)
+        sto_padding_item = TPDOMapItem(size_bits=cls.STO_COMMAND_PADDING_SIZE_BITS)
+        tpdo_map.add_item(sto_padding_item)
+        sbc_command_item = TPDOMapItem(size_bits=cls.SBC_COMMAND_SIZE_BITS)
+        tpdo_map.add_item(sbc_command_item)
+        safe_inputs_item = TPDOMapItem(size_bits=cls.SAFE_INPUTS_SIZE_BITS)
+        tpdo_map.add_item(safe_inputs_item)
+        safe_inputs_padding_item = TPDOMapItem(size_bits=cls.SAFE_INPUTS_PADDING_SIZE_BITS)
+        tpdo_map.add_item(safe_inputs_padding_item)
         crc_0_item = TPDOMapItem(size_bits=cls.CRC_O_SIZE_BITS)
         tpdo_map.add_item(crc_0_item)
         conn_id_item = TPDOMapItem(size_bits=cls.CONN_ID_SIZE_BITS)
@@ -240,12 +362,49 @@ class FSoEMaster:
     SAFETY_ADDRESS_REGISTER = EthercatRegister(
         idx=0x4193, subidx=0x00, dtype=REG_DTYPE.U16, access=REG_ACCESS.RW
     )
+    SAFE_INPUTS_MAP_REGISTER = EthercatRegister(
+        identifier="SAFE_INPUTS_MAP",
+        idx=0x46D2,
+        subidx=0x00,
+        dtype=REG_DTYPE.U16,
+        access=REG_ACCESS.RW,
+    )
+    STO_ACTIVATE_SBC_REGISTER = EthercatRegister(
+        identifier="STO_ACTIVATE_SBC",
+        idx=0x6643,
+        subidx=0x00,
+        dtype=REG_DTYPE.U32,
+        access=REG_ACCESS.RW,
+    )
+    SS1_TIME_TO_STO_REGISTER = EthercatRegister(
+        identifier="SS1_TIME_TO_STO",
+        idx=0x6651,
+        subidx=0x01,
+        dtype=REG_DTYPE.U16,
+        access=REG_ACCESS.RW,
+    )
+    SS1_ACTIVATE_SBC_REGISTER = EthercatRegister(
+        identifier="SS1_ACTIVATE_SBC",
+        idx=0x6658,
+        subidx=0x01,
+        dtype=REG_DTYPE.U32,
+        access=REG_ACCESS.RW,
+    )
+    SBC_BRAKE_TIME_DELAY_REGISTER = EthercatRegister(
+        identifier="SBC_BRAKE_TIME_DELAY",
+        idx=0x6661,
+        subidx=0x00,
+        dtype=REG_DTYPE.U16,
+        access=REG_ACCESS.RW,
+    )
 
     def __init__(self, motion_controller: "MotionController") -> None:
         self.logger = ingenialogger.get_logger(__name__)
         self.__mc = motion_controller
         self.__handlers: Dict[str, FSoEMasterHandler] = {}
         self.__next_connection_id = 1
+        self._error_observers: List[Callable[[FSoEError], None]] = []
+        self.__fsoe_configured = False
 
     def create_fsoe_master_handler(
         self,
@@ -260,8 +419,13 @@ class FSoEMaster:
 
         """
         slave_address = self.get_safety_address(servo)
+        application_parameters = self._get_application_parameters(servo)
         master_handler = FSoEMasterHandler(
-            slave_address, self.__next_connection_id, fsoe_master_watchdog_timeout
+            slave_address,
+            self.__next_connection_id,
+            fsoe_master_watchdog_timeout,
+            application_parameters,
+            partial(self._notify_errors, servo=servo),
         )
         self.__handlers[servo] = master_handler
         self.__next_connection_id += 1
@@ -278,6 +442,7 @@ class FSoEMaster:
         self._subscribe_to_pdo_thread_events()
         if start_pdos:
             self.__mc.capture.pdo.start_pdos()
+        self.__fsoe_configured = True
 
     def stop_master(self, stop_pdos: bool = False) -> None:
         """Stop all the FSoE Master handlers.
@@ -287,9 +452,14 @@ class FSoEMaster:
 
         """
         for master_handler in self.__handlers.values():
-            master_handler.stop()
-        self._unsubscribe_from_pdo_thread_events()
-        self._remove_pdo_maps_from_slaves()
+            if master_handler.running:
+                master_handler.stop()
+        if self.__fsoe_configured:
+            self._unsubscribe_from_pdo_thread_events()
+            self._remove_pdo_maps_from_slaves()
+        else:
+            self.logger.warning("FSoE master is already stopped")
+        self.__fsoe_configured = False
         if stop_pdos:
             self.__mc.capture.pdo.stop_pdos()
 
@@ -312,6 +482,39 @@ class FSoEMaster:
         """
         master_handler = self.__handlers[servo]
         master_handler.sto_activate()
+
+    def ss1_deactivate(self, servo: str = DEFAULT_SERVO) -> None:
+        """Deactivate the SS1.
+
+        Args:
+            servo: servo alias to reference it. ``default`` by default.
+
+        """
+        master_handler = self.__handlers[servo]
+        master_handler.ss1_deactivate()
+
+    def ss1_activate(self, servo: str = DEFAULT_SERVO) -> None:
+        """Activate the SS1.
+
+        Args:
+            servo: servo alias to reference it. ``default`` by default.
+
+        """
+        master_handler = self.__handlers[servo]
+        master_handler.ss1_activate()
+
+    def get_safety_inputs_value(self, servo: str = DEFAULT_SERVO) -> bool:
+        """Get a drive's safe inputs register value.
+
+        Args:
+            servo: servo alias to reference it. ``default`` by default.
+
+        Returns:
+           The safe inputs value.
+
+        """
+        master_handler = self.__handlers[servo]
+        return master_handler.safe_inputs_value()
 
     def get_safety_address(self, servo: str = DEFAULT_SERVO) -> int:
         """Get the drive's FSoE slave address.
@@ -367,6 +570,53 @@ class FSoEMaster:
         """
         master_handler = self.__handlers[servo]
         master_handler.wait_for_data_state(timeout)
+
+    def get_fsoe_master_state(self, servo: str = DEFAULT_SERVO) -> FSoEState:
+        """Get the servo's FSoE master handler state.
+
+        Args:
+            servo: servo alias to reference it. ``default`` by default.
+
+        Returns:
+            The servo's FSoE master handler state.
+
+        """
+        master_handler = self.__handlers[servo]
+        return master_handler.state
+
+    def subscribe_to_errors(self, callback: Callable[[FSoEError], None]) -> None:
+        """Subscribe to the FSoE errors.
+
+        Args:
+            callback: Subscribed callback function.
+
+        """
+        if callback in self._error_observers:
+            return
+        self._error_observers.append(callback)
+
+    def unsubscribe_from_errors(self, callback: Callable[[FSoEError], None]) -> None:
+        """Unsubscribe from the FSoE errors.
+
+        Args:
+            callback: Subscribed callback function.
+
+        """
+        if callback not in self._error_observers:
+            return
+        self._error_observers.remove(callback)
+
+    def _notify_errors(self, transition_name: str, error_description: str, servo: str) -> None:
+        """Notify subscribers when an FSoE error occurs.
+
+        Args:
+            transition_name: FSoE transition name.
+            error_description: FSoE error description.
+            servo: The servo alias.
+
+        """
+        for callback in self._error_observers:
+            callback(FSoEError(servo, transition_name, error_description))
 
     def _delete_master_handler(self, servo: str = DEFAULT_SERVO) -> None:
         """Delete the master handler instance
@@ -427,3 +677,23 @@ class FSoEMaster:
             "The FSoE Master lost connection to the FSoE slaves. "
             f"An exception occurred during the PDO exchange: {exc}"
         )
+
+    def _get_application_parameters(self, servo: str) -> List["ApplicationParameter"]:
+        """Get values of the application parameters"""
+        drive = self.__mc.servos[servo]
+        application_parameters = []
+        for register in [
+            self.SAFE_INPUTS_MAP_REGISTER,
+            self.STO_ACTIVATE_SBC_REGISTER,
+            self.SS1_TIME_TO_STO_REGISTER,
+            self.SS1_ACTIVATE_SBC_REGISTER,
+            self.SBC_BRAKE_TIME_DELAY_REGISTER,
+        ]:
+            register_size_bytes, _ = dtype_value[register.dtype]
+            application_parameter = ApplicationParameter(
+                name=register.identifier,
+                initial_value=drive.read(register),
+                n_bytes=register_size_bytes,
+            )
+            application_parameters.append(application_parameter)
+        return application_parameters