+++ ingeniamotion/wizard_tests/feedbacks_tests/absolute_encoder2_test.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, List
+from typing import TYPE_CHECKING, List, Optional
 
 if TYPE_CHECKING:
     from ingeniamotion import MotionController
@@ -14,8 +14,10 @@ class AbsoluteEncoder2Test(Feedbacks):
 
     SENSOR_TYPE_FEEDBACK_TEST = SensorType.BISSC2
 
-    def __init__(self, mc: "MotionController", servo: str, axis: int):
-        super().__init__(mc, servo, axis)
+    def __init__(
+        self, mc: "MotionController", servo: str, axis: int, logger_drive_name: Optional[str] = None
+    ) -> None:
+        super().__init__(mc, servo, axis, logger_drive_name)
         self.backup_registers_names.extend(self.BACKUP_REGISTERS_BISSC2)
 
     @BaseTest.stoppable